Added AboutUs, Rules, fixed last bugs, etc

+++ API/SignalR/GameLobbyHub.cs
@@ -1,46 +1,34 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using API.Entities;
 using API.Extensions;
 using API.Interfaces;
-using API.SignalIR;
-using AutoMapper;
 using Microsoft.AspNetCore.SignalR;
 
 namespace API.SignalR
-{
-    // incomplete
+{    
     public class GameLobbyHub : Hub
     {
-        private readonly IMapper _mapper;
-        private readonly IHubContext<PresenceHub> _presenceHub;
-        private readonly PresenceTracker _tracker;
         private readonly IUnitOfWork _unitOfWork;
 
-        public GameLobbyHub(IUnitOfWork unitOfWork, IMapper mapper,
-            IHubContext<PresenceHub> presenceHub, PresenceTracker tracker)
+        public GameLobbyHub(
+            IUnitOfWork unitOfWork)
         {
             _unitOfWork = unitOfWork;
-            _tracker = tracker;
-            _presenceHub = presenceHub;
-            _mapper = mapper;
         }
 
         public override async Task OnConnectedAsync()
         {
             var httpContext = Context.GetHttpContext();
             var strgameLobbyId = httpContext.Request.Query["lobbyId"].ToString();
-            var gameLobbyId = Int32.Parse(strgameLobbyId);
-            var guest = Context.User;
+            var gameLobbyId = Int32.Parse(strgameLobbyId);           
 
             // later check if it works 
             //var gameLobbies = GetLobbies();
             //await Clients.Caller.SendAsync("GetGameLobbies", gameLobbies);
 
-            var gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            var gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            //if (gameLobby != null) throw new HubException("No Game Available");
+
             var groupName = gameLobby.GameLobbyName;                 
             await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
 
@@ -57,7 +45,10 @@ namespace API.SignalR
         public override async Task OnDisconnectedAsync(Exception exception)
         {
             var group = await RemoveFromLobby();
-            await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
+            if (group != null)
+            {
+                await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
+            }            
             await base.OnDisconnectedAsync(exception);
         }    
 
@@ -73,13 +64,12 @@ namespace API.SignalR
 
         private async Task<Group> AddToLobby(string groupName, GameLobby gameLobby)
         {
-            //var connection = new Connection(Context.ConnectionId, Context.User.GetUsername()/*, game*/);
-            var connection = new Connection();
+            Connection connection;
 
             // case where the user disconnects
             if (gameLobby.GameStatus == "ongoing")
             {
-                connection = await _unitOfWork.ConnectionRepository.GetConnection(Context.User.GetUsername());
+                connection = _unitOfWork.ConnectionRepository.GetConnection(Context.User.GetUsername());
                 if (connection == null) throw new HubException("This game already started!");
                 connection.ConnectionId = Context.ConnectionId;
 
@@ -110,13 +100,14 @@ namespace API.SignalR
 
         private async Task<Group> RemoveFromLobby()
         {
-            var group = await _unitOfWork.GameLobbyRepository.GetGroupForConnection(Context.ConnectionId);            
+            var group = await _unitOfWork.GameLobbyRepository.GetGroupForConnection(Context.ConnectionId);
+            if (group == null) return group;
 
             if (group.Connections.Count > 0)
             {
-                var gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyByName(group.Name);
+                var gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyByName(group.Name);
                 var connection = group.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
-                await _unitOfWork.GameLobbyRepository.RemoveConnection(gameLobby, group, connection);
+                _unitOfWork.GameLobbyRepository.RemoveConnection(gameLobby, group, connection);
             }          
             
             if (await _unitOfWork.Complete()) return group;
@@ -128,13 +119,13 @@ namespace API.SignalR
         {
             var httpContext = Context.GetHttpContext();
             var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());
-            var gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            var gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
             var group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);
 
             if (privateRoom) 
             {
                 StringBuilder builder = new StringBuilder();
-                Random random = new Random();
+                Random random = new();
                 char ch;
                 for (int i = 0; i < 10; i++)
                 {
@@ -145,7 +136,7 @@ namespace API.SignalR
                 gameLobby.Password = builder.ToString();
             } else
             {
-                gameLobby.Password = " ";
+                gameLobby.Password = "";
             }
 
             if (await _unitOfWork.Complete())
@@ -157,17 +148,16 @@ namespace API.SignalR
             {
                 throw new HubException("Failed to make it private/public the game!");
             }
-        }
+        }        
         
-        // Currently working on this
         public async Task<GameLobby> StartGame()
         {
             var httpContext = Context.GetHttpContext();
-            var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());            
+            var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());
 
-            var gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
-            if (gameLobby == null) throw new HubException("That game lobby does not exist!");            
-            
+            var gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            if (gameLobby == null) throw new HubException("That game lobby does not exist!");
+
             if (gameLobby.NumberOfElements < 2) throw new HubException("Waiting for more players");
 
             if (gameLobby.GameStatus == "ongoing") throw new HubException("The game has already started.");
@@ -191,14 +181,13 @@ namespace API.SignalR
         {
             var httpContext = Context.GetHttpContext();
             var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());
-            //var gameLobbyId = Int32.Parse(strgameLobbyId);
             var username = Context.User.GetUsername();
 
-            GameLobby gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            GameLobby gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
             if (gameLobby.CurrentPlayer != username) throw new HubException("It is " + gameLobby.CurrentPlayer + "'s turn, it is not your turn!");
            
             var group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);
-            Connection connection = await _unitOfWork.ConnectionRepository.GetConnection(username);
+            Connection connection = _unitOfWork.ConnectionRepository.GetConnection(username);
 
             string message = await _unitOfWork.GameLobbyRepository.Play(connection, gameLobby, cards);
             if (message != "Next") throw new HubException(message);
@@ -206,7 +195,7 @@ namespace API.SignalR
             if (cards.First().Value == -1)
             {
                 int i = 0;
-                while (i < cards.Count())
+                while (i < cards.Count)
                 {
                     message = await _unitOfWork.GameLobbyRepository.GetConsequence(gameLobby, group, connection, cards);
                     if (message == "Card type is incorrect!") throw new HubException(message);
@@ -214,16 +203,16 @@ namespace API.SignalR
                 }
             }   
 
-            var numberOfCards = connection.Cards.Count();
+            var numberOfCards = connection.Cards.Count;
             if  (numberOfCards == 0)
             {
+                gameLobby.Winner = gameLobby.CurrentPlayer;
                 gameLobby.GameStatus = "finished";
             }
 
-            if (message != "Pick a colour")
+            if (message != "Pick a colour" && gameLobby.GameStatus != "finished")
             {
-                // get the next turn
-                var turn = await _unitOfWork.GameLobbyRepository.NextTurn(gameLobby, group);
+                var turn = _unitOfWork.GameLobbyRepository.NextTurn(gameLobby, group);
                 if (!turn) throw new HubException("I cannot get to the next turn!");
             }
 
@@ -244,13 +233,13 @@ namespace API.SignalR
             var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());
             var username = Context.User.GetUsername();
 
-            GameLobby gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            GameLobby gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
             if (gameLobby.CurrentPlayer != username) return "It is " + gameLobby.CurrentPlayer + "'s turn, it is not your turn!";
 
             var group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);
-            Connection connection = await _unitOfWork.ConnectionRepository.GetConnection(username);
+            Connection connection = _unitOfWork.ConnectionRepository.GetConnection(username);
 
-            var message = await _unitOfWork.GameLobbyRepository.PlayWithChosenColour(connection, gameLobby, cards, colour);
+            var message = _unitOfWork.GameLobbyRepository.PlayWithChosenColour(connection, gameLobby, cards, colour);
             if (message != "Next") throw new HubException(message);
            
             gameLobby.PickedColour = "none";
@@ -258,7 +247,7 @@ namespace API.SignalR
             if (cards.First().Value == -1)
             {
                 int i = 0;
-                while (i < cards.Count())
+                while (i < cards.Count)
                 {
                     message = await _unitOfWork.GameLobbyRepository.GetConsequence(gameLobby, group, connection, cards);
                     if (message == "Card type is incorrect!") throw new HubException(message);
@@ -266,10 +255,16 @@ namespace API.SignalR
                 }
             }
 
-            if (message != "Pick a colour")
+            var numberOfCards = connection.Cards.Count;
+            if (numberOfCards == 0)
             {
-                // get the next turn
-                var turn = await _unitOfWork.GameLobbyRepository.NextTurn(gameLobby, group);
+                gameLobby.Winner = gameLobby.CurrentPlayer;
+                gameLobby.GameStatus = "finished";
+            }
+
+            if (message != "Pick a colour" && gameLobby.GameStatus != "finished")
+            {
+                var turn = _unitOfWork.GameLobbyRepository.NextTurn(gameLobby, group);
                 if (!turn) throw new HubException("I cannot get to the next turn!");
             }
 
@@ -291,16 +286,21 @@ namespace API.SignalR
             var httpContext = Context.GetHttpContext();
             var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());
 
-            bool validate = await _unitOfWork.GameLobbyRepository.PickColour(colour);
+            bool validate = _unitOfWork.GameLobbyRepository.PickColour(colour);
             if (!validate) throw new HubException("Colour is not valid!");
 
-            GameLobby gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            GameLobby gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
             gameLobby.PickedColour = colour;
 
             var group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);
-
-            bool turn = await _unitOfWork.GameLobbyRepository.NextTurn(gameLobby, group);
+            
+            bool turn = _unitOfWork.GameLobbyRepository.NextTurn(gameLobby, group);
             if (!turn) throw new HubException("It is not possible get to the next turn!");
+
+            if (gameLobby.FileName == "bWildDraw4")
+            {
+                turn = _unitOfWork.GameLobbyRepository.NextTurn(gameLobby, group);
+            }
 
             if (await _unitOfWork.Complete())
             {
@@ -316,22 +316,21 @@ namespace API.SignalR
             var httpContext = Context.GetHttpContext();
             var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());
 
-            GameLobby gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
-            var group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);
-            Connection connection = await _unitOfWork.ConnectionRepository.GetConnection(gameLobby.CurrentPlayer);                      
+            GameLobby gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            Group group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);
+            Connection connection = _unitOfWork.ConnectionRepository.GetConnection(gameLobby.CurrentPlayer);                      
 
             Card pot = await _unitOfWork.CardRepository.GetCard(gameLobby.LastCard);
             ICollection<Card> cards = connection.Cards;
 
-            // verify if the current player have a valid card to play
-            bool playable = await _unitOfWork.GameLobbyRepository.Playable(pot, cards);
+            // verify if the current player has a valid card to play
+            bool playable = _unitOfWork.GameLobbyRepository.Playable(pot, cards);
             if (playable) return "You have cards that you can play!";
 
             if (connection.Uno == true) connection.Uno = false;
 
-            // get a card from deck until we can play
-            Card cardFromDeck = new Card();
-            // working - test with more cases
+            // get a card from deck until the player has 'playable' card
+            Card cardFromDeck;
             do
             {
                 cardFromDeck = await _unitOfWork.GameLobbyRepository.Draw(1, gameLobby, connection);
@@ -343,7 +342,6 @@ namespace API.SignalR
 
             if (await _unitOfWork.Complete())
             {
-
                 await Clients.Group(group.Name).SendAsync("GetGameLobby", gameLobby);
                 await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
                 return "You obtained the cards required!";
@@ -359,19 +357,20 @@ namespace API.SignalR
             var httpContext = Context.GetHttpContext();
             var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());
 
-            GameLobby gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            GameLobby gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
             var group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);
-            Connection connection = await _unitOfWork.ConnectionRepository.GetConnection(gameLobby.CurrentPlayer);      
+            Connection connection = _unitOfWork.ConnectionRepository.GetConnection(gameLobby.CurrentPlayer);      
 
             ICollection<Card> cards = connection.Cards;
 
-            // verify if the current player have a valid card to play
-            bool playable = await _unitOfWork.GameLobbyRepository.PlayableWithColour(cards, colour);
+            // verify if the current player has a valid card to play
+            bool playable = _unitOfWork.GameLobbyRepository.PlayableWithColour(cards, colour);
             if (playable) return "You have cards that you can play!";
 
             if (connection.Uno == true) connection.Uno = false;
-            // get a card from deck until we can play
-            Card cardFromDeck = new Card();
+
+            // get a card from deck until the player has 'playable' card
+            Card cardFromDeck;
             do
             {
                 cardFromDeck = await _unitOfWork.GameLobbyRepository.Draw(1, gameLobby, connection);
@@ -397,13 +396,12 @@ namespace API.SignalR
             var httpContext = Context.GetHttpContext();
             var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());
 
-            GameLobby gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            GameLobby gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
             var group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);               
             
-            Connection connection = await _unitOfWork.ConnectionRepository.GetConnection(Context.User.GetUsername());
-            //if (gameLobby.CurrentPlayer != connection.Username) throw new HubException("It is not your turn!");
+            Connection connection = _unitOfWork.ConnectionRepository.GetConnection(Context.User.GetUsername());
 
-            var message = await _unitOfWork.GameLobbyRepository.UnoStatus(connection);
+            var message = _unitOfWork.GameLobbyRepository.UnoStatus(connection);
             if (message == "You cannot say uno!") throw new HubException(message);
                        
             if (await _unitOfWork.Complete())
@@ -420,12 +418,12 @@ namespace API.SignalR
 
         public async Task<string> CatchUno(string username)
         {
-            Connection connection = await _unitOfWork.ConnectionRepository.GetConnection(username);
+            Connection connection = _unitOfWork.ConnectionRepository.GetConnection(username);
             if (connection.Uno == true) return "Too late! " + connection.Username + " has already said UNO!";
 
             var httpContext = Context.GetHttpContext();
             var gameLobbyId = Int32.Parse(httpContext.Request.Query["lobbyId"].ToString());
-            GameLobby gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            GameLobby gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
             var group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);
 
             await _unitOfWork.GameLobbyRepository.Draw(4, gameLobby, connection);
@@ -450,7 +448,7 @@ namespace API.SignalR
             var msg = _unitOfWork.GameLobbyRepository.DeleteGame(gameLobbyId);
             if (msg == "Connection does not exist") return msg;
 
-            GameLobby gameLobby = await _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
+            GameLobby gameLobby = _unitOfWork.GameLobbyRepository.GetGameLobbyById(gameLobbyId);
             Group group = await _unitOfWork.GameLobbyRepository.GetGroup(gameLobby.GameLobbyName);
 
             foreach (Connection connection in group.Connections)
